pyproject.toml
[project]
name = "pdfsummarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.23.0",
    "litellm>=1.67.1",
    "loguru>=0.7.3",
    "openai>=1.75.0",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.0",
    "streamlit>=1.44.1",
]
